E2E scheduling of pods to GW/NonGW nodes
Before this patch, E2E didn't have any control of where pods
landed, this made the testing of different scenarios random.

With this patch we ensure the testing of the datapath when
traffic needs to transit from a worker node to a gateway node.

Cluster type has been converted to an enumeration as part of this
commit.