Fix network unreachable issue on AWS with multiple AZs
When deploying a K8s/OpenShift Cluster on AWS, multiple Availability Zones (AZs) are
used for maximized service availability. Each AZ uses a CIDR (f.e., 10.165.32.0/20)
thats pulled out of the VPC Global CIDR (f.e., 10.165.0.0/16).

Currently in Submariner, we are deriving the VxLAN VTEP IP address from the host-ip as
shown below.

For example: Say, the host ipaddress is "192.168.1.100/16", we were prepending 240 to
the host-ip address and configuring the vxlan vtepIP as "240.168.1.100/16"

This works fine when all the nodes are on the same Subnet CIDR, but will have issues
when programming inter-cluster routing rules on the host via the vx-submariner interface.

Error message:
route.go:719] Error adding route {Ifindex: 58 Dst: 10.252.0.0/14 Src: <nil>
 Gw: 240.165.41.54 Flags: [] Table: 0}: network is unreachable

This patch modifies the implementation to use vtepIP Mask of 8 (i.e., 240.168.1.100/8)
so that when adding inter-cluster routing rules, the nextHop (i.e., remoteVTEPIP) is on
the same network and routing rules can be programmed successfully.