Fix race condition causing NO_PROPOSAL_CHOSEN
When one side of the cluster installs a cable a connection request is
automatically triggered. If the other side's `charon` has not been
configured yet with the correct IKE then it answers with a NO_PROPOSAL_CHOSEN
causing the initiator side to remove its SA and not retry again. This
leads to a situation where the cable is installed, and thus won't be
replaced/reinstalled, but without an SA to establish the tunnel. A
"stale" cable of sorts.

This problem becomes more relevant and apparent when only one side has a
public IP, as in this case the system cannot rely on the reverse SA to
establish the tunnel.

This patch handles the scenario where a cable becomes stale by removing
that cable and allowing submariner to install it again, reestablishing
a valid SA.