Remove the operator sourcecode from this repository. (#190)
We could as well remove the operator deployment scripts at some
point, we'd need to refactor CI moving the operator deploy logic
somewhere else. Same thing for helm.

But that needs a bigger change.

For now, e2e run in this repo will be pinpointed to 0.0.1 of the
operator and we can change that later.