Use Endpoint CRD to configure routes
Currently the `submariner-agent` makes use of `Cluster` CRDs to set
up the routes. This, however, is information that is more natural to
the `Endpoint` CRD as it describes the subnets that are to be found
when connecting that endpoint. Moreover, having a single source of
truth instead of two simplifies the code and allows for easier
integration with `submariner-globalnet`.

This patch makes the `Endpoint` be the source of truth for setting up
routes. By doing so, there is no need for the `submariner-routeagent` to
monitor `Cluster`s anymore and therefore the code that is related has
been removed, simplifying the codebase in consequence.