Use globalCIDR from ClusterCRD
Currently, globalCIDR is configured/stored as part of clusterCRD and
Submariner uses this value to populate the endpoint subnets. At the
same time, globalnet controller is reading this value as an environment
parameter in the globalnet POD. Storing/reading globalCIDR from two
different places could be error prone if not configured/deployed properly.
This patch modifies globalnet controller to read the globalCIDR from the
clusterCRD which is the source of truth for a cluster.

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>