Move connection types to the submariner.io API
We want to expose the connection status in the operator status. The
initial plan was to make the status available over HTTP, and query
that from the operator; however, the operator only updates its status
in the reconciliation loop, and apart from that the status is static.

This means that operator status updates require watching Kubernetes
objects which reflect the state encapsulated in the operator status.
To enable this, we can expose the status types in Submariner alongside
the existing Cluster and Endpoint types; to avoid package import
loops, we move everything there, and adjust the existing cable engine
and drivers accordingly.

Signed-off-by: Stephen Kitt <skitt@redhat.com>