libreswan: HostNetworking on GW to remote cluster support
This PR implements the necessary support in route-agent code to enable
remote cluster connectivity from a Pod on GatewayNode with HostNetworking
for libreswan cable driver.

Along with this, it also fixes an issue with Strongswan (inc' globalnet)
where the sourceIP of such traffic was not properly configured to the
CNIInterfaceIP of the GatewayNode.

Fixes issue: https://github.com/submariner-io/submariner/issues/635

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>