Fix shutdown path in Submariner Gateway Pod
In the current code, it was seen that GatewayPod was not handling the
shutdown signal properly and was finally getting terminated by K8s
after the terminationGracePeriod interval is exhaused. This is because
leaderelection.RunOrDie does not return. This PR fixes it by running
the leaderelection in a Goroutine and using stopCh to wait for shutdown
signal.

Partially fixes: https://github.com/submariner-io/submariner/issues/694

Signed-Off-by: Sridhar Gaddam <sgaddam@redhat.com>