Fix intermittent event controller test failures
A race condition with the 'registry' field:

Write at 0x00c0003a3fa0 by goroutine 9:
  github.com/submariner-io/submariner/pkg/event/controller_test...
    /go/src/github.com/submariner-io/submariner/pkg/event/
       controller/controller_test.go:43

Previous read at 0x00c0003a3fa0 by goroutine 21:
  github.com/submariner-io/submariner/pkg/event.(*Registry).invokeHandlers()
      /go/src/github.com/submariner-io/submariner/pkg/event/registry.go:141
  github.com/submariner-io/submariner/pkg/event.(*Registry).StopHandlers()
      /go/src/github.com/submariner-io/submariner/pkg/event/registry.go:67

This is due to the test starting the controller in a Go routine but not
waiting for it to complete. There's anogther intermittent failure
related to fake client sets/informers that is also due to the async Start.

To fix both I changed Start to not wait on the stopCh. This also aligns
with the semantics of Start vs Run.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>