Reset gateway connection status when health checker reports connected
When the health checker reports a connection error, the gateway syncer
sets the gateway connection status to error. When the health checker
subsequently reports connected, the gateway syncer needs to set the
status back to connected.

The rest of the changes were related to adding unit tests for the health
checker's interaction with the gateway syncer. This entailed creating
a PingerInterface and a fake implementation that can be injected into the
health checker.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>