Associate submariner_router to chassis via options
In OVN, you can provide a list of "gateway_chassis" for a router port along
with priorities,  or you can simply attach the router to a chassis.

The former works more reliably in our case. And we don't need the extra features
of "Gateway_Chassis" so this flips the implementation to the router
options field.

Signed-off-by: Miguel Angel Ajo <majopela@redhat.com>