Notify NAT discovery RemoveEndpoint in cable engine
On RemoveCable, we need to call RemoveEndpoint on the NAT
discovery.

Due to the async nature of install cable, an edge case is if
RemoveCable occurs while NAT discovery is in progress. There's a
timing window where the cable could be installed after RemoveCable.
Added data structures and logic to prevent this. Also
driver.DisconnectFromEndpoint should not be called if cable installation
did not successfully complete.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>