Handle retries on install cable failure
Fixes https://github.com/submariner-io/submariner/issues/1243

The simple solution implemented is to set a resync interval in the
tunnel controller. 30 sec was chosen as a reasonable interval as we don't
expect failures to be common. This also means that every endpoint will be
resynced every 30 sec but it should essentially be a no-op if the endpoint
was already successfully installed. Prior to the change to use the resource
watcher, we had a resync interval set anyway (60 sec).

Modified natDiscovery.AddEndpoint to notify the ready channel if discovery
had already completed. This enables retries to immediately  proceed again to
the install cable phase.

The tunnel controller tests were changed to use the real cable engine and
NAT discovery to more effectively test the retries round trip.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>