Ignore CNI Interface when healthcheck is disabled
On some platforms, it was seen that CNI Interface on the node
was missing. Submariner uses the CNIInterface IP for certain use-cases
like HostNetwork to remoteCluster support and for Healthcheck use-cases.
The current code was expecting the CNIInterface to be present even
when the healthcheck is disabled, this PR fixes it.

Signed-Off-by: Sridhar Gaddam <sgaddam@redhat.com>