Add GHA for SubM Tweeting via PRs
Add a GitHub action that will enable the community to manage
Submariner's Twitter by PRs to .tweet files.

This can be used for announcing releases, as Submariner's Twitter is
currently mostly used for, but can also be used for other Submariner
content. For example, to highlight media about Submariner or events
Submariner will take part in. In general, this opens up Submariner's
Twitter to the community, subject to the community's code review.

This also increases Submariner's "bus factor", which is a requirement
for CII Silver.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>