Build Submariner images for x86-64 and ARM64
This relies on cross-building support in the Shipyard base image and
project Makefiles to avoid having to provide a native Shipyard base
image (and slower compilation using QEMU). The final image build stage
relies on support for the target architecture however (for now).

The TARGETPLATFORM default is dropped from the container definitions;
it is expected to be provided by the build environment (buildx
currently), and setting it prevents us from actually building foreign
images.

Docker architectures are translated to Fedora architectures for the
networkplugin-syncer image, to retrieve the appropriate ovn package.

Signed-off-by: Stephen Kitt <skitt@redhat.com>